Logic and communication between controller working for user login
Still need to figure out how to pass string from jsp to java code.  Also I had to hard code some of the connection stuff in my userauthentication function.  The generic results set kept returning a null type for me so I wasn't able to use it in this case.