Added DataModification Class
DataModification holds functions for adding data, updating data, and deleting data. Note that updateRowString is for updating columns with any datatype that isn't an int, like a name or date, while updateRowVal updates columns with int or bigint datatypes, like id numbers or years.

Functions to add entries to maintenance, incoming, and outgoing have yet to be written.